Check HTTP status 204 on file uploads
Indicates that although files were uploaded, their content was identical to what was already on the server, and so nothing was changed.

This can be further improved by returning a boolean indicating whether something changed or not, and using that to determine whether a repo sync is necessary before uploading more files.